Decouple JobBootstrap with OneOffJobBootstrap and ScheduleJobBootstrap
